Use thumbprint instead of friendly name for certificate labels.
Using the friendly name property (CERT_FRIENDLY_NAME_PROP_ID) as the CKA_LABEL
attribute value for a certificate can be problematic, as not all certificates
have a value for this property. This has been seen to cause problems with NSS,
where multiple certificates without friendly names were assigned the same
nickname. This resulted in a browser only being able to access one of the
certificates.

This fix updates p11c_cert_certificate_get_bytes() to report the certificate's
thumbprint instead (provided by the CERT_HASH_PROP_ID property). Thumbprints
are commonly used to identify certificates on Windows systems, and can be
calculated for any certificate (as they are simply a hash).

Signed-off-by: Simon Haggett <simon.haggett@gmail.com>